requirements.txt
Django==4.2.7
python-decouple==3.8
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-allauth==0.58.2
django-filter==23.3
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
django-tailwind==3.8.0
python-dotenv==1.0.0
dj-database-url==2.1.0
django-storages==1.14.2
sentry-sdk==1.32.0
django-cors-headers==4.3.0
django-debug-toolbar==4.2.0
django-csp==3.7
django-redis==5.4.0
